// Capstone Project/src/app/pages/user/suggestion-page/screen/types.ts
export enum searchPageEnum {
  GET_HOUSE_POSTING_REQUEST = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_REQUEST',
  GET_HOUSE_POSTING_SUCCESS = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_SUCCESS',
  GET_HOUSE_POSTING_FAILURE = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_FAILURE',

  GET_HOUSE_POSTING_LOAD_MORE_REQUEST = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_LOAD_MORE_REQUEST',
  GET_HOUSE_POSTING_LOAD_MORE_SUCCESS = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_LOAD_MORE_SUCCESS',
  GET_HOUSE_POSTING_LOAD_MORE_FAILURE = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_LOAD_MORE_FAILURE',

  GET_HOUSE_POSTING_SUGGESTION_REQUEST = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_SUGGESTION_REQUEST',
  GET_HOUSE_POSTING_SUGGESTION_SUCCESS = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_SUGGESTION_SUCCESS',
  GET_HOUSE_POSTING_SUGGESTION_FAILURE = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_SUGGESTION_FAILURE',

  ADD_TO_FAVOURITE_REQUEST = 'APP/SEARCH_PAGE/ADD_TO_FAVOURITE_REQUEST',
  ADD_TO_FAVOURITE_SUCCESS = 'APP/SEARCH_PAGE/ADD_TO_FAVOURITE_SUCCESS',
  ADD_TO_FAVOURITE_FAILURE = 'APP/SEARCH_PAGE/ADD_TO_FAVOURITE_FAILURE',

  GET_POST_TOP_8_REQUEST = 'APP/SEARCH_PAGE/GET_POST_TOP_8_REQUEST',
  GET_POST_TOP_8_SUCCESS = 'APP/SEARCH_PAGE/GET_POST_TOP_8_SUCCESS',
  GET_POST_TOP_8_FAILURE = 'APP/SEARCH_PAGE/GET_POST_TOP_8_FAILURE',

  GET_HOUSE_POSTING_MAP_REQUEST = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_MAP_REQUEST',
  GET_HOUSE_POSTING_MAP_SUCCESS = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_MAP_SUCCESS',
  GET_HOUSE_POSTING_MAP_FAILURE = 'APP/SEARCH_PAGE/GET_HOUSE_POSTING_MAP_FAILURE',

  FILTER_HOUSE_POSTING_REQUEST = 'APP/SEARCH_PAGE/FILTER_HOUSE_POSTING_REQUEST',
  FILTER_HOUSE_POSTING_SUCCESS = 'APP/SEARCH_PAGE/FILTER_HOUSE_POSTING_SUCCESS',
  FILTER_HOUSE_POSTING_FAILURE = 'APP/SEARCH_PAGE/FILTER_HOUSE_POSTING_FAILURE',

  FILTER_HOUSE_POSTING_MAP_REQUEST = 'APP/SEARCH_PAGE/FILTER_HOUSE_POSTING_MAP_REQUEST',
  FILTER_HOUSE_POSTING_MAP_SUCCESS = 'APP/SEARCH_PAGE/FILTER_HOUSE_POSTING_MAP_SUCCESS',
  FILTER_HOUSE_POSTING_MAP_FAILURE = 'APP/SEARCH_PAGE/FILTER_HOUSE_POSTING_MAP_FAILURE',

  CLEAR_SEARCH_PAGE_STATE = 'APP/SEARCH_PAGE/CLEAR_SEARCH_PAGE_STATE',
}